requirements.txt
# Core dependencies
python-dotenv>=1.0.0
pydantic>=2.0.0
typer>=0.9.0
fastapi>=0.95.0
uvicorn>=0.15.0

# PDF Processing
PyPDF2>=3.0.0
pdf2image>=1.16.0
pytesseract>=0.3.10
google-cloud-documentai>=2.20.0

# NLP & Text Processing
nltk>=3.8.1
transformers>=4.30.0
sentence-transformers>=2.2.2

# Machine Learning
torch>=2.0.0
accelerate>=0.20.0
peft>=0.5.0
datasets>=2.12.0
evaluate>=0.4.0
rouge-score>=0.1.2
matplotlib>=3.7.0
optuna>=3.0.0
seaborn>=0.12.0

# Google Cloud & Vertex AI
google-cloud-storage>=2.10.0
google-cloud-aiplatform>=1.30.0
google-cloud-monitoring>=2.0.0
google-cloud-logging>=3.0.0

# Pipeline & Orchestration
kfp>=2.0.0

# Testing
pytest>=7.3.1
pytest-cov>=4.1.0

# Development
black>=23.3.0
isort>=5.12.0
mypy>=1.3.0
flake8>=6.0.0
jupyter>=1.0.0
